requirements.txt
Flask==3.1.2
pymongo
dnspython
python-dotenv
google-generativeai
